feat: dummy chart integrated with profile
