+++ package.json
@@ -14,10 +14,12 @@
     "@devexpress/dx-react-core": "^4.0.0",
     "@emotion/react": "^11.10.5",
     "@emotion/styled": "^11.10.5",
+    "@faker-js/faker": "^7.6.0",
     "@mui/icons-material": "^5.10.16",
     "@mui/joy": "^5.0.0-alpha.56",
     "@mui/material": "^5.10.16",
     "@next/font": "^13.0.6",
+    "@types/chart.js": "^2.9.37",
     "@types/node": "18.11.10",
     "@types/react-dom": "18.0.9",
     "axios": "^1.2.1",
